Hide ArborX::PairIndexRank
With APIv2, it is only used in tests. Unless we decide to expose it to a
user, but it may be pretty inconvenient, and I don't currently see a way
to easily produce such result without callbacks.
